improve wx-downloader (#67)
* improve wx-downloader

* refine code

+++ wechatgame/libs/wx-downloader.js
@@ -31,7 +31,7 @@ var non_text_format = [
 
 var REGEX = /^\w+:\/\/.*/;
 
-var fs = wx.getFileSystemManager ? wx.getFileSystemManager() : {};
+var fs = wx.getFileSystemManager ? wx.getFileSystemManager() : null;
 
 var _newAssets = [];
 var WXDownloader = window.WXDownloader = function () {
@@ -47,8 +47,7 @@ WXDownloader.ID = ID;
 WXDownloader.prototype.handle = function (item, callback) {
 
     if (item.type === 'js') {
-        callback(null, null);
-        return;
+        return null;
     }
     if (item.type === 'uuid') {
         var result = cc.Pipeline.Downloader.PackDownloader.load(item, callback);
@@ -66,8 +65,7 @@ WXDownloader.prototype.handle = function (item, callback) {
 
     if (CC_WECHATGAMESUB) {
         if (REGEX.test(item.url)) {
-            callback(null, null);
-            return;
+            return null;
         }
 
         item.url = this.SUBCONTEXT_ROOT + '/' + item.url;
@@ -75,6 +73,11 @@ WXDownloader.prototype.handle = function (item, callback) {
         if (item.type && non_text_format.indexOf(item.type) !== -1) {
             nextPipe(item, callback);
             return;
+        }
+
+        // if wx.getFileSystemManager is undefined, need to skip
+        if (!fs) {
+            return null;
         }
     }
 
@@ -111,6 +114,10 @@ WXDownloader.prototype.cleanOldAssets = function () {
 };
 
 function cleanAllFiles(path, newAssets, finish) {
+    if (!fs) {
+        finish('wx.getFileSystemManager is undefined');
+        return;
+    }
     fs.readdir({
         dirPath: path,
         success: function (res) {
@@ -145,13 +152,13 @@ function cleanAllFiles(path, newAssets, finish) {
                     }
                 }
                 else {
-                    finish && finish();
+                    finish();
                 }
 
             })(0);
         },
         fail: function (res) {
-            finish && finish();
+            finish(res ? res.errMsg : 'unknown error');
         },
     });
 }
@@ -174,7 +181,7 @@ function nextPipe(item, callback) {
     var queue = cc.LoadingItems.getQueue(item);
     queue.addListener(item.id, function (item) {
         if (item.error) {
-            fs.unlink({
+            fs && fs.unlink({
                 filePath: item.url,
                 success: function () {
                     cc.log('Load failed, removed local file ' + item.url + ' successfully!');
@@ -186,6 +193,13 @@ function nextPipe(item, callback) {
 }
 
 function readText (item, callback) {
+    if (!fs) {
+        callback({
+            status: 0,
+            errorMessage: 'wx.getFileSystemManager is undefined'
+        });
+        return;
+    }
     var url = item.url;
     fs.readFile({
         filePath: url,
@@ -231,6 +245,14 @@ function readText (item, callback) {
 }
 
 function readFromLocal (item, callback) {
+    if (!fs) {
+        callback({
+            status: 0,
+            errorMessage: 'wx.getFileSystemManager is undefined'
+        });
+        return;
+    }
+
     var localPath = wx.env.USER_DATA_PATH + '/' + item.url;
 
     // Read from local file cache
@@ -262,6 +284,11 @@ function readFromLocal (item, callback) {
 }
 
 function ensureDirFor (path, callback) {
+    if (!fs) {
+        callback('wx.getFileSystemManager is undefined');
+        return;
+    }
+
     // cc.log('mkdir:' + path);
     var ensureDir = cc.path.dirname(path);
     if (ensureDir === "wxfile://usr" || ensureDir === "http://usr") {
